# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Furniture Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: string;
  name: string;
  category: string;
  price: number;
  description: string;
  dimensions: string;
  material: string;
  image: string;
  thumbnails: string[];
  createdAt?: string;
}

export interface OrderItem {
  productId: string;
  name: string;
  price: number;
  quantity: number;
}

export interface PlaceOrderRequest {
  /** @minLength 1 */
  customerName: string;
  customerEmail: string;
  /** @minLength 1 */
  shippingAddress: string;
  /** @minItems 1 */
  items: OrderItem[];
  subtotal: number;
  shippingCost: number;
  tax: number;
  total: number;
  paymentMethod?: string;
}

export interface PlaceOrderResponse {
  success: boolean;
  orderNumber: string;
  orderId: number;
}

export interface CartSyncItem {
  productId: string;
  quantity: number;
  priceAtTime: number;
}

export interface CartSyncRequest {
  items: CartSyncItem[];
}

export interface CartSyncResponse {
  success: boolean;
}

export interface ContactMessage {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  message: string;
}

export interface ContactResponse {
  success: boolean;
  message: string;
}

export interface PublicFaqItem {
  id: number;
  question: string;
  answer: string;
  isFeatured: boolean;
  createdAt: string;
}

export type AdminFaqItemStatus =
  (typeof AdminFaqItemStatus)[keyof typeof AdminFaqItemStatus];

export const AdminFaqItemStatus = {
  pending: "pending",
  answered: "answered",
} as const;

export interface AdminFaqItem {
  id: number;
  productId: string;
  customerName: string;
  customerEmail: string;
  question: string;
  answer?: string | null;
  status: AdminFaqItemStatus;
  isFeatured: boolean;
  createdAt: string;
}

export interface SubmitFaqRequest {
  /** @minLength 1 */
  productId: string;
  /** @minLength 1 */
  customerName: string;
  customerEmail: string;
  /** @minLength 10 */
  question: string;
}

export interface SubmitFaqResponse {
  success: boolean;
  message: string;
  id: number;
}

export interface AnswerFaqRequest {
  /** @minLength 1 */
  answer: string;
  isFeatured?: boolean;
}

export interface ErrorResponse {
  error: string;
}

export type ListAdminFaqsParams = {
  status?: ListAdminFaqsStatus;
};

export type ListAdminFaqsStatus =
  (typeof ListAdminFaqsStatus)[keyof typeof ListAdminFaqsStatus];

export const ListAdminFaqsStatus = {
  pending: "pending",
  answered: "answered",
} as const;