# lib/api-zod/src/generated/types/adminFaqItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Furniture Store API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminFaqItemStatus } from "./adminFaqItemStatus";

export interface AdminFaqItem {
  id: number;
  productId: string;
  customerName: string;
  customerEmail: string;
  question: string;
  answer?: string | null;
  status: AdminFaqItemStatus;
  isFeatured: boolean;
  createdAt: Date;
}