# lib/api-zod/src/generated/types/cartSyncRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Furniture Store API
 * OpenAPI spec version: 0.1.0
 */
import type { CartSyncItem } from "./cartSyncItem";

export interface CartSyncRequest {
  items: CartSyncItem[];
}