# lib/api-zod/src/generated/types/contactResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Furniture Store API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactResponse {
  success: boolean;
  message: string;
}