# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Furniture Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: string;
  name: string;
  category: string;
  price: number;
  description: string;
  dimensions: string;
  material: string;
  image: string;
  thumbnails: string[];
  createdAt?: Date;
}